+++ app/javascript/vue/components/radials/linker/constants/links.js
@@ -95,3 +95,10 @@ export const TASK_DWC_OCCURRENCE_STATUS = {
   label: 'DwC Occurrence Status',
   link: '/tasks/dwc_occurrences/status'
 }
+
+export const TASK_NEW_CONTAINER = {
+  label: 'New container',
+  link: '/tasks/containers/new_container',
+  queryParam: true,
+  saveQuery: true
+}