+++ app/javascript/vue/composables/index.js
@@ -1,5 +1,6 @@
 export * from './useBeforeUnloadListener'
 export * from './useDraggable'
+export * from './useInterval'
 export * from './useOnResize'
 export * from './usePopstateListener'
 export * from './useScroll'