+++ app/javascript/vue/tasks/containers/new_container/constants/index.js
@@ -1,0 +1,3 @@
+export * from './colors'
+export * from './options'
+export * from './parameters.js'