+++ app/javascript/vue/tasks/containers/new_container/constants/options.js
@@ -1,0 +1,13 @@
+export const DEFAULT_OPTS = {
+  enclose: true,
+  itemSize: 1,
+  padding: 1,
+  canvas: {
+    backgroundColor: '#0f172a'
+  },
+  cameraPosition: {
+    x: 50,
+    y: 50,
+    z: 50
+  }
+}