+++ app/models/collection_object.rb
@@ -115,6 +115,7 @@ class CollectionObject < ApplicationRecord
   # Repository delegations
   delegate :acronym, to: :repository, prefix: :repository, allow_nil: true
   delegate :url, to: :repository, prefix: :repository, allow_nil: true
+  delegate :institutional_LSID, to: :repository, prefix: :repository, allow_nil: true
 
   # Preparation delegations
   delegate :name, to: :preparation_type, prefix: :preparation_type, allow_nil: true
@@ -631,6 +632,7 @@ class CollectionObject < ApplicationRecord
     h
   end
 
+  # TODO: Unify with Extract in concern
   # @return [Identifier::Local::CatalogNumber, nil]
   #   the first (position) catalog number for this collection object, either on specimen, or container
   def preferred_catalog_number