+++ spec/models/depiction_spec.rb
@@ -5,6 +5,13 @@ RSpec.describe Depiction, type: :model, groups: [:images, :observation_matrix] d
   let(:image_file) { Rack::Test::UploadedFile.new( Spec::Support::Utilities::Files.generate_png, 'image/png') }
   let(:specimen) { FactoryBot.create(:valid_specimen) }
 
+  specify 'dwc_occurrence hooks' do
+    s = Specimen.create!
+    expect(s.dwc_occurrence.associatedMedia).to eq(nil)
+    d = Depiction.create!(depiction_object: s, image: FactoryBot.create(:valid_image))
+    expect(s.dwc_occurrence.reload.associatedMedia).to match('http://127.0.0.1:3000/api/v1/images/')
+  end
+
   specify 'when moving to a new collection object old is destroyed if an image stub' do
     o1 = FactoryBot.create(:valid_collection_object)
     o2 = FactoryBot.create(:valid_collection_object)
@@ -12,7 +19,7 @@ RSpec.describe Depiction, type: :model, groups: [:images, :observation_matrix] d
     d = FactoryBot.create(:valid_depiction, depiction_object: o2)
 
     FactoryBot.create(:valid_sqed_depiction, depiction: d)
-    
+
     d.update!(depiction_object: o1)
 
     expect(CollectionObject.where(id: o2.id).any?).to be_falsey
@@ -20,7 +27,7 @@ RSpec.describe Depiction, type: :model, groups: [:images, :observation_matrix] d
 
   specify 'destroying depiction does not destroy related Observation::Media 2 when not last' do
     o = FactoryBot.create(:valid_observation, type: 'Observation::Media', descriptor: Descriptor::Media.create!(name: 'test'))
-    
+
     d = FactoryBot.create(:valid_depiction, depiction_object: o)
 
     # Can not test with this because of the way callbacks are created
@@ -46,7 +53,7 @@ RSpec.describe Depiction, type: :model, groups: [:images, :observation_matrix] d
     o = FactoryBot.create(:valid_observation, type: 'Observation::Media', descriptor: Descriptor::Media.create!(name: 'test'))
     d1 = FactoryBot.create(:valid_depiction, depiction_object: o)
 
-    otu = Otu.create!(name: 'Foo') 
+    otu = Otu.create!(name: 'Foo')
 
     d2 = Depiction.create(depiction_object: otu, image: d1.image)